refactor(embeddings): split text sidecar store
